Fixes: #12853 Move tests from pg dialect
